updated index and status endpoints

+++ src/main/java/net/englab/contextsearcher/services/VideoIndexer.java
@@ -8,16 +8,23 @@ import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import net.englab.contextsearcher.elastic.VideoDocument;
 import net.englab.contextsearcher.models.EnglishVariety;
+import net.englab.contextsearcher.elastic.IndexMetadata;
+import net.englab.contextsearcher.models.IndexingInfo;
 import net.englab.contextsearcher.models.SrtSentence;
 import net.englab.contextsearcher.models.entities.Video;
 import net.englab.contextsearcher.utils.SrtSentenceParser;
+import org.springframework.http.HttpStatus;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
+import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.Future;
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.stream.Collectors;
 
 import static net.englab.contextsearcher.elastic.ElasticProperties.*;
@@ -28,12 +35,30 @@ import static net.englab.contextsearcher.elastic.ElasticProperties.*;
 public class VideoIndexer {
 
     private final static int BULK_SIZE = 10_000;
-    private final static String VIDEOS_INDEX = "videos";
+    public final static String VIDEOS_INDEX = "videos";
 
     private final VideoStorage videoStorage;
     private final ElasticService elasticService;
 
+    private final ThreadPoolTaskExecutor executor;
+    private IndexingInfo indexingInfo = IndexingInfo.none();
+    private final AtomicBoolean isRunning = new AtomicBoolean(false);
+
+    /**
+     * Adds a new video.
+     *
+     * @param videoId   the video id
+     * @param variety   the variety of English used in the video
+     * @param srt       the subtitles for the video in SRT format
+     * @param index     true if we also want to index the video
+     */
     public void add(String videoId, EnglishVariety variety, String srt, boolean index) {
+        if (index && isRunning.get()) {
+            throw new ResponseStatusException(
+                    HttpStatus.CONFLICT,
+                    "A video cannot be indexed while an indexing job is running"
+            );
+        }
         Video video = new Video(null, videoId, variety, srt);
         Long id = videoStorage.save(video);
         if (!index) return;
@@ -46,8 +71,20 @@ public class VideoIndexer {
         }
     }
 
+    /**
+     * Removes the given video.
+     *
+     * @param videoId   the video id
+     * @param index     true if we also want to remove the video from the index
+     */
     public void remove(String videoId, boolean index) {
         if (index) {
+            if (isRunning.get()) {
+                throw new ResponseStatusException(
+                        HttpStatus.CONFLICT,
+                        "A video cannot be removed from the index while an indexing job is running"
+                );
+            }
             try {
                 elasticService.removeVideo(VIDEOS_INDEX, videoId);
             } catch (Exception e) {
@@ -58,21 +95,60 @@ public class VideoIndexer {
         videoStorage.deleteByVideoId(videoId);
     }
 
-    public void indexAll() {
-        log.info("Full indexing has been started.");
-        elasticService.removeIndex(VIDEOS_INDEX);
-        log.info("The old index has been removed.");
-        indexVideos(videoStorage.findAll());
+    /**
+     * Returns the current status of indexing.
+     *
+     * @return indexing info
+     */
+    public IndexingInfo getIndexingStatus() {
+        if (!isRunning.get()) {
+            elasticService.getIndexMetadata(VIDEOS_INDEX)
+                    .ifPresent(meta -> indexingInfo = IndexingInfo.completed(meta.startTime(), meta.finishTime()));
+        }
+        return indexingInfo;
+    }
+
+    /**
+     * Starts a new indexing job.
+     */
+    public void startIndexing() {
+        if (isRunning.getAndSet(true)) {
+            throw new ResponseStatusException(
+                    HttpStatus.CONFLICT,
+                    "A new indexing job cannot be started if one is already running"
+            );
+        };
+        indexingInfo = IndexingInfo.started(Instant.now());
+        executor.execute(() -> {
+            try {
+                log.info("Full indexing has been started.");
+                elasticService.removeIndex(VIDEOS_INDEX);
+                log.info("The old index has been removed.");
+                log.info("Start reading videos from the database...");
+                List<Video> videos = videoStorage.findAll();
+                log.info("Start indexing the videos...");
+                indexVideos(videos);
+                log.info("Indexing has been finished successfully.");
+            } catch (Throwable throwable) {
+                indexingInfo = IndexingInfo.failed(indexingInfo.startTime(), Instant.now(), throwable.getMessage());
+                log.error("An exception occurred during indexing", throwable);
+                throw new RuntimeException(throwable);
+            } finally {
+                isRunning.set(false);
+            }
+        });
     }
 
     @SneakyThrows
     private void indexVideos(Collection<Video> videos) {
+        Instant startTime = Instant.now();
         elasticService.createIndexIfAbsent(VIDEOS_INDEX, Map.of(
                 "video_id", KEYWORD_PROPERTY,
                 "sentence", TEXT_PROPERTY,
                 "variety", KEYWORD_PROPERTY,
                 "subtitle_blocks", ObjectProperty.of(b -> b.enabled(false))._toProperty()
         ));
+        log.info("A new index has been created.");
         List<Future<BulkResponse>> futures = bulkIndex(videos);
         for (Future<BulkResponse> future : futures) {
             BulkResponse response = future.get();
@@ -82,6 +158,8 @@ public class VideoIndexer {
                 log.info("{} docs have been successfully indexed. It took {}ms.", response.items().size(), response.took());
             }
         }
+        elasticService.setIndexMetadata(VIDEOS_INDEX, new IndexMetadata(startTime, Instant.now()));
+        log.info("The index metadata has been updated.");
     }
 
     private List<Future<BulkResponse>> bulkIndex(Collection<Video> videos) {
@@ -108,7 +186,7 @@ public class VideoIndexer {
         return futures;
     }
 
-    public Map<String, Integer> rangesToMap(RangeMap<Integer, Integer> ranges) {
+    private static Map<String, Integer> rangesToMap(RangeMap<Integer, Integer> ranges) {
         return ranges.asMapOfRanges().entrySet().stream()
                 .collect(Collectors.toMap(e -> e.getKey().toString(), Map.Entry::getValue));
     }
